Added Game Manager and level changing
Added game manager for handling level transitions and made it and the player static entities which will persist between level transitions
Note: each level should have the Player object and a PlayerStart prefab. Placing a game manager isnt necessary as long as the first scene in the game has it.